백업 : 동기화를 위한 임시 Static용 클래스 생성, Card - SetPositionAndRotation에 동기화 파라미터 추가

+++ Assets/UdonScript/HandCalculator.cs
@@ -1459,7 +1459,7 @@ public class HandCalculator : UdonSharpBehaviour
     Card TEST__SetTestData(Card card, string type, int cardNumber)
     {
         card.Type = type;
-        card.CardNumber = cardNumber;
+        card.CardNumber = (uint)cardNumber;
         card.GlobalOrder = HandUtil.GetGlobalOrder(type, cardNumber);
         return card;
     }
